requirements.txt
# Function dependencies, for example:
# package>=version

Flask
line-bot-sdk
requests
google-generativeai
git+https://github.com/ozgur/python-firebase